Fixed issue #444 by renaming testsuite_test_client directory.
Initial commit, Broken into modules but no real test yet

Breaks listener and subscription manager apart. Adds at least some
support for static filters, subscriptions.

modify test code, clean up

1. Removed some code from _listener that is in subscription manager

2. Corrected listener_id processing in subscription manager

3. Corrected listener_url in subscription manager

4. Completed the examples peg indication test. This test now runs

5. Completed changes to run_cimoperations.py for new interfaces
