+++ pywbem/_listener.py
@@ -106,7 +106,6 @@ from six.moves import socketserver
 from six.moves import http_client
 
 from . import cim_xml
-from ._server import WBEMServer
 from ._version import __version__
 from .cim_obj import CIMInstance, CIMInstanceName, _ensure_unicode
 from .cim_operations import check_utf8_xml_chars
@@ -603,6 +602,7 @@ class WBEMListener(object):
     def __init__(self, host, http_port=DEFAULT_LISTENER_PORT_HTTP,
                  https_port=DEFAULT_LISTENER_PORT_HTTPS,
                  certfile=None, keyfile=None, listener_id=None):
+## TODO eliminate listener_id from here unless there is a need
         """
         Parameters:
 
@@ -719,13 +719,9 @@ class WBEMListener(object):
         with all attributes, that is suitable for debugging.
         """
         return "%s(host=%r, http_port=%s, https_port=%s, " \
-               "certfile=%r, keyfile=%r, logger=%r, _servers=%r, " \
-               "_subscription_paths=%r, _dynamic_filter_paths=%r, " \
-               "_destination_path=%r, _callbacks=%r)" % \
+               "certfile=%r, keyfile=%r, logger=%r, _callbacks=%r)" % \
                (self.__class__.__name__, self.host, self.http_port,
                 self.https_port, self.certfile, self.keyfile, self.logger,
-                self._servers, self._subscription_paths,
-                self._dynamic_filter_paths, self._destination_path,
                 self._callbacks)
 
     @property
@@ -882,397 +878,6 @@ class WBEMListener(object):
             self._https_server = None
             self._https_thread = None
 
-    def add_server(self, server):
-        """
-        Add a WBEM server to the listener and register the listener with the
-        server by creating an indication listener instance referencing this
-        listener in the Interop namespace of the server.
-
-        Parameters:
-
-          server (:class:`~pywbem.WBEMServer`):
-            The WBEM server.
-
-        Returns:
-
-            :term:`string`: An ID for the WBEM server, for use by other
-            methods of this class.
-
-        Raises:
-
-            Exceptions raised by :class:`~pywbem.WBEMConnection`.
-        """
-
-        if not isinstance(server, WBEMServer):
-            raise TypeError("server argument of add_server() must be a " \
-                            "WBEMServer object")
-        server_id = server.url
-        if server_id in self._servers:
-            raise ValueError("WBEM server already known by listener: %s" % \
-                             server_id)
-
-        # We let the WBEM server use HTTP or HTTPS dependent on whether we
-        # contact it using HTTP or HTTPS.
-        if server.conn.url.lower().startswith('http'):
-            scheme = 'http'
-            port = self.http_port
-        elif server.conn.url.lower().startswith('https'):
-            scheme = 'https'
-            port = self.https_port
-        else:
-            raise ValueError("Invalid scheme in server URL: %s" % \
-                             server.conn.url)
-
-        dest_url = '%s://%s:%s' % (scheme, self.host, port)
-        dest_path = _create_destination(server, dest_url, self._listener_id)
-
-        self._servers[server_id] = server
-        self._subscription_paths[server_id] = []
-        self._dynamic_filter_paths[server_id] = []
-        self._destination_path[server_id] = dest_path
-
-        return server_id
-
-    def remove_server(self, server_id):
-        """
-        Remove a WBEM server from the listener and unregister the listener
-        from the server by deleting all indication subscriptions, dynamic
-        indication filters and listener destinations in the server that were
-        created by this listener.
-
-        Parameters:
-
-          server_id (:term:`string`):
-            The server ID for the WBEM server, returned by :meth:`add_server`.
-
-        Raises:
-
-            Exceptions raised by :class:`~pywbem.WBEMConnection`.
-        """
-
-        if server_id not in self._servers:
-            raise ValueError("WBEM server not known by listener: %s" % \
-                             server_id)
-        server = self._servers[server_id]
-
-        # Delete any instances we recorded to be cleaned up
-
-        if server_id in self._subscription_paths:
-            paths = self._subscription_paths[server_id]
-            for i, path in enumerate(paths):
-                server.conn.DeleteInstance(path)
-            del self._subscription_paths[server_id]
-
-        if server_id in self._dynamic_filter_paths:
-            paths = self._dynamic_filter_paths[server_id]
-            for i, path in enumerate(paths):
-                server.conn.DeleteInstance(path)
-            del self._dynamic_filter_paths[server_id]
-
-        if server_id in self._destination_path:
-            path = self._destination_path[server_id]
-            server.conn.DeleteInstance(path)
-            del self._destination_path[server_id]
-
-        # Remove server from this listener
-        del self._servers[server_id]
-
-    def add_dynamic_filter(self, server_id, source_namespace, query,
-                           query_language=DEFAULT_QUERY_LANGUAGE,
-                           filter_id=None):
-        """
-        Add a dynamic indication filter to a WBEM server, by creating an
-        indication filter instance in the Interop namespace of the server.
-
-        Dynamic indication filters are those that are created by clients.
-        Indication filters that pre-exist in the WBEM server are termed
-        *static indication filters* and cannot be created or deleted by
-        clients. See :term:`DSP1054` for details about indication filters.
-
-        Parameters:
-
-          server_id (:term:`string`):
-            The server ID for the WBEM server, returned by :meth:`add_server`.
-
-          source_namespace (:term:`string`):
-            Source namespace of the indication filter.
-
-          query (:term:`string`):
-            Filter query in the specified query language.
-
-          query_language (:term:`string`):
-            Query language for the specified filter query.
-
-            Examples: 'WQL', 'DMTF:CQL'.
-
-        filter_id(:term: `string`)
-            If not `None` a printable string that is incorporated into the name
-            property of the filter instance to help the listener identify
-            its filters and handler instances in a WBEM Server. The character
-            ":" is not allowed in this string since it is the separator between
-            components of the final filter_id.  Filter ids need not be
-            unique and can be used to identify groups of filters by applying
-            the same `filter_id` to them.
-
-            The form of the filter name property is:
-
-              "pywbemfilter:" [ <listener_id> ":"] [ <filter_id> ":" ] <guid>
-
-
-        Returns:
-
-            :class:`~pywbem.CIMInstanceName`: The instance path of the
-            indication filter instance.
-
-        Raises:
-
-            Exceptions raised by :class:`~pywbem.WBEMConnection`.
-        """
-        if server_id not in self._servers:
-            raise ValueError("WBEM server not known by listener: %s" % \
-                             server_id)
-        server = self._servers[server_id]
-        filter_path = _create_filter(server, source_namespace, query,
-                                     query_language, self._listener_id,
-                                     filter_id)
-        self._dynamic_filter_paths[server_id].append(filter_path)
-        return filter_path
-
-    def remove_dynamic_filter(self, server_id, filter_path):
-        """
-        Remove a dynamic indication filter from a WBEM server, by deleting an
-        indication filter instance in the WBEM server.
-
-        The indication filter must be a dynamic indication filter that has been
-        created by this listener, and there must not exist any subscriptions
-        referencing the filter.
-
-        Parameters:
-
-          server_id (:term:`string`):
-            The server ID for the WBEM server, returned by :meth:`add_server`.
-
-          filter_path (:class:`~pywbem.CIMInstanceName`):
-            Instance path of the indication filter instance in the WBEM
-            server.
-
-        Raises:
-
-            Exceptions raised by :class:`~pywbem.WBEMConnection`.
-        """
-        if server_id not in self._servers:
-            raise ValueError("WBEM server not known by listener: %s" % \
-                             server_id)
-        server = self._servers[server_id]
-        if server_id in self._dynamic_filter_paths:
-            paths = self._dynamic_filter_paths[server_id]
-            for i, path in enumerate(paths):
-                if path == filter_path:
-                    server.conn.DeleteInstance(path)
-                    del paths[i]
-                    break
-
-    def get_dynamic_filters(self, server_id):
-        """
-        Return the dynamic indication filter instance paths in a WBEM server
-        that have been created by this listener. This function access only
-        the local list of dynamic filters. It does NOT contact the WBEM
-        server.
-
-        Parameters:
-
-          server_id (:term:`string`):
-            The server ID for the WBEM server, returned by :meth:`add_server`.
-
-        Returns:
-
-            List of :class:`~pywbem.CIMInstanceName`: The dynamic indication
-            filter instance paths.
-
-        Raises:
-
-            Exceptions raised by :class:`~pywbem.WBEMConnection`.
-        """
-        if server_id not in self._servers:
-            raise ValueError("WBEM server not known by listener: %s" % \
-                             server_id)
-        return self._dynamic_filter_paths[server_id]
-
-    def get_filters(self, server_id):
-        """
-        Return all (dynamic and static) indication filter instance paths in
-        a WBEM server. This function requests filters from the WBEM server
-
-        Parameters:
-
-          server_id (:term:`string`):
-            The server ID for the WBEM server, returned by :meth:`add_server`.
-
-        Returns:
-
-          List of :class:`~pywbem.CIMInstanceName`: The indication filter
-          instance paths.
-
-        Raises:
-
-            Exceptions raised by :class:`~pywbem.WBEMConnection`.
-        """
-        if server_id not in self._servers:
-            raise ValueError("WBEM server not known by listener: %s" % \
-                             server_id)
-        server = self._servers[server_id]
-        return server.conn.EnumerateInstanceNames('CIM_IndicationFilter',
-                                                  namespace=server.interop_ns)
-
-    def add_subscription(self, server_id, filter_path):
-        """
-        Add a subscription to a WBEM server for particular set of indications
-        defined by an indication filter, by creating an indication subscription
-        instance in the Interop namespace of the server, that links the
-        specified indication filter with the listener destination in the
-        server that references this listener.
-
-        The indication filter can be a dynamic filter created specifically for
-        this listener via :meth:`add_dynamic_filter`, or a static filter that
-        pre-exists in the WBEM server. Filters defined in the WBEM server can
-        be retrieved via :meth:`get_filters`.
-
-        Upon successful return of this method, the subscription is active.
-
-        Parameters:
-
-          server_id (:term:`string`):
-            The server ID for the WBEM server, returned by :meth:`add_server`.
-
-          filter_path (:class:`~pywbem.CIMInstanceName`):
-            Instance path of the indication filter instance in the WBEM
-            server that specifies the indications to be sent.
-
-        Returns:
-
-            :class:`~pywbem.CIMInstanceName`: Instance path of the indication
-            subscription instance.
-
-        Raises:
-
-            Exceptions raised by :class:`~pywbem.WBEMConnection`.
-        """
-        if server_id not in self._servers:
-            raise ValueError("WBEM server not known by listener: %s" % \
-                             server_id)
-        server = self._servers[server_id]
-        dest_path = self._destination_path[server_id]
-        sub_path = _create_subscription(server, dest_path, filter_path)
-        self._subscription_paths[server_id].append(sub_path)
-        return sub_path
-
-    def remove_subscription(self, server_id, sub_path):
-        """
-        Remove an indication subscription from a WBEM server, by deleting an
-        indication subscription instance in the server.
-
-        Parameters:
-
-          server_id (:term:`string`):
-            The server ID for the WBEM server, returned by :meth:`add_server`.
-
-          sub_path (:class:`~pywbem.CIMInstanceName`):
-            Instance path of the indication subscription instance in the WBEM
-            server.
-
-        Raises:
-
-            Exceptions raised by :class:`~pywbem.WBEMConnection`.
-        """
-        if server_id not in self._servers:
-            raise ValueError("WBEM server not known by listener: %s" % \
-                             server_id)
-        server = self._servers[server_id]
-        if server_id in self._subscription_paths:
-            sub_paths = self._subscription_paths[server_id]
-            for i, path in enumerate(sub_paths):
-                if path == sub_path:
-                    server.conn.DeleteInstance(path)
-                    del sub_paths[i]
-                    break
-
-    def get_subscriptions(self, server_id):
-        """
-        Return the indication subscriptions in a WBEM server that have been
-        created by this listener.
-
-        Parameters:
-
-          server_id (:term:`string`):
-            The server ID for the WBEM server, returned by :meth:`add_server`.
-
-        Returns:
-
-            List of :class:`~pywbem.CIMInstanceName`: The indication
-            subscription instance paths.
-
-        Raises:
-
-            Exceptions raised by :class:`~pywbem.WBEMConnection`.
-        """
-        if server_id not in self._servers:
-            raise ValueError("WBEM server not known by listener: %s" % \
-                             server_id)
-        return self._subscription_paths[server_id]
-
-    def get_all_destination_instances(self, server_id):
-        """
-        Return all 'CIM_ListenerDestinationCIMXML' instance paths in a
-        WBEM server.
-
-        Parameters:
-
-          server_id (:term:`string`):
-            The server ID for the WBEM server, returned by :meth:`add_server`.
-
-        Returns:
-
-          List of :class:`~pywbem.CIMInstanceName`: The
-          CIM_ListenerDestinationCIMXML instance paths.
-
-        Raises:
-
-            Exceptions raised by :class:`~pywbem.WBEMConnection`.
-        """
-
-        if server_id not in self._servers:
-            raise ValueError("WBEM server not known by listener: %s" % \
-                             server_id)
-        server = self._servers[server_id]
-        return server.conn.EnumerateInstanceNames(DESTINATION_CLASSNAME,
-                                                  namespace=server.interop_ns)
-
-    def get_all_subscriptions(self, server_id):
-        """
-        Return all CIM_IndicationSubscription instance paths in a WBEM server.
-
-        Parameters:
-
-          server_id (:term:`string`):
-            The server ID for the WBEM server, returned by :meth:`add_server`.
-
-        Returns:
-
-          List of :class:`~pywbem.CIMInstanceName`: The subscription
-          instance paths.
-
-        Raises:
-
-            Exceptions raised by :class:`~pywbem.WBEMConnection`.
-        """
-
-        if server_id not in self._servers:
-            raise ValueError("WBEM server not known by listener: %s" % \
-                             server_id)
-        server = self._servers[server_id]
-        return server.conn.EnumerateInstanceNames(SUBSCRIPTION_CLASSNAME,
-                                                  namespace=server.interop_ns)
-
     def _deliver_indication(self, indication, host):
         """
         This function is called by the listener threads for each received
@@ -1310,175 +915,6 @@ class WBEMListener(object):
         """
         if callback not in self._callbacks:
             self._callbacks.append(callback)
-
-
-def _create_destination(server, dest_url, listener_id=None):
-    """
-    Create a listener destination instance in the Interop namespace of a
-    WBEM server and return its instance path.
-
-    Parameters:
-
-      server (:class:`~pywbem.WBEMServer`):
-        Identifies the WBEM server.
-
-      dest_url (:term:`string`):
-        URL of the listener that is used by the WBEM server to send any
-        indications to.
-
-        The URL scheme (e.g. http/https) determines whether the WBEM server
-        uses HTTP or HTTPS for sending the indication. Host and port in the
-        URL specify the target location to be used by the WBEM server.
-
-    listener_id (:term: `string`)
-        If not `None` a string of printiable characters that prefixes the name
-        property of filters and listener destination instances to help the
-        listener identify its filters and handler instances in a WBEM Server.
-        This id is used as a component of the `Name` property of the
-        destination instance.
-
-    Returns:
-
-        :class:`~pywbem.CIMInstanceName`: The instance path of the created
-        instance.
-
-    Raises:
-
-        Exceptions raised by :class:`~pywbem.WBEMConnection`.
-    """
-
-    dest_path = CIMInstanceName(DESTINATION_CLASSNAME)
-    dest_path.classname = DESTINATION_CLASSNAME
-    dest_path.namespace = server.interop_ns
-
-    dest_inst = CIMInstance(DESTINATION_CLASSNAME)
-    dest_inst.path = dest_path
-    dest_inst['CreationClassName'] = DESTINATION_CLASSNAME
-    dest_inst['SystemCreationClassName'] = SYSTEM_CREATION_CLASSNAME
-    dest_inst['SystemName'] = getfqdn()
-    dest_inst['Name'] = 'pywbemlistener:%s%s' % (listener_id, uuid.uuid4())
-    dest_inst['Destination'] = dest_url
-
-    dest_path = server.conn.CreateInstance(dest_inst)
-    return dest_path
-
-def _create_filter(server, source_namespace, query, query_language, \
-                   listener_id=None, filter_id=None):
-    """
-    Create a dynamic indication filter instance in the Interop namespace
-    of a WBEM server and return its instance path.
-
-    Parameters:
-
-      server (:class:`~pywbem.WBEMServer`):
-        Identifies the WBEM server.
-
-      source_namespace (:term:`string`):
-        Source namespace of the indication filter.
-
-      query (:term:`string`):
-        Filter query in the specified query language.
-
-      query_language (:term:`string`):
-        Query language for the specified filter query.
-
-        Examples: 'WQL', 'DMTF:CQL'.
-
-      filter_id (:term:`string`):
-        String prefex that will be prepended to the created filter. Can
-        be used to help identify filters in the server. If `None` no prefix
-        is added to the system-defined name for this filter
-
-      listener_id (:term: `string`))
-        If not `None` a string of printable characters to be inserted in the
-        name property of filter and listener destination instances to help the
-        listener identify its filters and destination instances in a
-        WBEM Server. The string must not contain the ":" character.
-
-        The form for the name property of a PyWBEM of a filter is:
-
-        "pywbemfilter:" [<listener_id>  ":"] [<filter_id> ":"] <guid>
-
-    Returns:
-
-        :class:`~pywbem.CIMInstanceName`: The instance path of the created
-        instance.
-
-    Raises:
-
-        Exceptions raised by :class:`~pywbem.WBEMConnection`.
-    """
-
-    filter_path = CIMInstanceName(FILTER_CLASSNAME)
-    filter_path.classname = FILTER_CLASSNAME
-    filter_path.namespace = server.interop_ns
-
-    filter_inst = CIMInstance(FILTER_CLASSNAME)
-    if filter_id is None:
-        filter_id_ = ''
-    elif isinstance(filter_id, six.string_types):
-        if ':' in filter_id:
-            raise ValueError("Invalid String: filter_id %s contains ':'" % \
-                             filter_id)
-        filter_id_ = '%s:' % filter_id
-    else:
-        raise TypeError("invalid type for filter_id=%s" % filter_id)
-
-    filter_inst.path = filter_path
-    filter_inst['CreationClassName'] = FILTER_CLASSNAME
-    filter_inst['SystemCreationClassName'] = SYSTEM_CREATION_CLASSNAME
-    filter_inst['SystemName'] = getfqdn()
-    filter_inst['Name'] = 'pywbemfilter:%s%s%s' % (listener_id, \
-                                                   filter_id_, \
-                                                   uuid.uuid4())
-    filter_inst['SourceNamespace'] = source_namespace
-    filter_inst['Query'] = query
-    filter_inst['QueryLanguage'] = query_language
-
-    filter_path = server.conn.CreateInstance(filter_inst)
-    return filter_path
-
-def _create_subscription(server, dest_path, filter_path):
-    """
-    Create an indication subscription instance in the Interop namespace of
-    a WBEM server and return its instance path.
-
-    Parameters:
-
-      server (:class:`~pywbem.WBEMServer`):
-        Identifies the WBEM server.
-
-      dest_path (:class:`~pywbem.CIMInstanceName`):
-        Instance path of the listener destination instance in the WBEM
-        server that references this listener.
-
-      filter_path (:class:`~pywbem.CIMInstanceName`):
-        Instance path of the indication filter instance in the WBEM
-        server that specifies the indications to be sent.
-
-    Returns:
-
-        :class:`~pywbem.CIMInstanceName`: The instance path of the created
-        instance.
-
-    Raises:
-
-        Exceptions raised by :class:`~pywbem.WBEMConnection`.
-    """
-
-
-    sub_path = CIMInstanceName(SUBSCRIPTION_CLASSNAME)
-    sub_path.classname = SUBSCRIPTION_CLASSNAME
-    sub_path.namespace = server.interop_ns
-
-    sub_inst = CIMInstance(SUBSCRIPTION_CLASSNAME)
-    sub_inst.path = sub_path
-    sub_inst['Filter'] = filter_path
-    sub_inst['Handler'] = dest_path
-
-    sub_path = server.conn.CreateInstance(sub_inst)
-    return sub_path
-
 
 def callback_interface(indication, host):
     # pylint: disable=unused-argument
