+++ testsuite/testclient/embeddedobjects.yaml
