+++ testsuite/testclient/getalltypes.yaml
